requirements.txt
langchain 
wikipedia 
langchain-community 
langchain_core
langchain_classic
langchain_groq
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
rich